refactor: use u8 instead of usize in Node for s1_choice and s2_choice
this saves a moderate amount of memory, maybe 5% of total. Casting to usize when needed is basically free.

+++ src/mcts.rs
@@ -23,8 +23,8 @@ pub struct Node {
 
     // represents the instructions & s1/s2 moves that led to this node from the parent
     pub instructions: StateInstructions,
-    pub s1_choice: usize,
-    pub s2_choice: usize,
+    pub s1_choice: u8,
+    pub s2_choice: u8,
 
     // represents the total score and number of visits for this node
     // de-coupled for s1 and s2
@@ -132,8 +132,8 @@ impl Node {
             let mut new_node = Node::new(s1_options, s2_options);
             new_node.parent = self;
             new_node.instructions = state_instructions;
-            new_node.s1_choice = s1_move_index;
-            new_node.s2_choice = s2_move_index;
+            new_node.s1_choice = s1_move_index as u8;
+            new_node.s2_choice = s2_move_index as u8;
 
             this_pair_vec.push(new_node);
         }
@@ -153,11 +153,11 @@ impl Node {
             return;
         }
 
-        let parent_s1_movenode = &mut (*self.parent).s1_options[self.s1_choice];
+        let parent_s1_movenode = &mut (*self.parent).s1_options[self.s1_choice as usize];
         parent_s1_movenode.total_score += score;
         parent_s1_movenode.visits += 1;
 
-        let parent_s2_movenode = &mut (*self.parent).s2_options[self.s2_choice];
+        let parent_s2_movenode = &mut (*self.parent).s2_options[self.s2_choice as usize];
         parent_s2_movenode.total_score += 1.0 - score;
         parent_s2_movenode.visits += 1;
